# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMP Stock Watcher API Documentations",
        "contact": {
            "email": "i.g.b.n.satyawibawa@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://warulaga.dapurkode.com",
            "description": "Dapurkode API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Testing Server"
        }
    ],
    "paths": {
...
        "/api/wallet/get-resource-crypto": {
            "get": {
                "tags": [
                    "Holdlist Crypto Controller",
                    "Holdlist Get Resource"
                ],
                "summary": "Cryptocurrencies that avaiable to add on holdlist",
                "operationId": "holdlistCryptoGetResource",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Searching keyword",
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collections of cryptoResources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "cryptoResources": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Id of collection",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "symbol": {
                                                        "description": "Symbol of stock",
                                                        "type": "string",
                                                        "example": "TSLA"
                                                    },
                                                    "name": {
                                                        "description": "Name of stock",
                                                        "type": "string",
                                                        "example": "Tesla Inc"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
        "/api/wallet/store-crypto": {
            "post": {
                "tags": [
                    "Holdlist Crypto Controller",
                    "Holdlist Store"
                ],
                "summary": "Add cryptocurrency to holdlist",
                "operationId": "holdlistCryptoStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/api/wallet/remove-crypto": {
            "post": {
                "tags": [
                    "Holdlist Crypto Controller",
                    "Holdlist Remove"
                ],
                "summary": "Remove crypto from holdlist",
                "operationId": "holdlistCryptoRemove",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}